requirements.txt
openai
supabase
pypdf2
python-dotenv
